[test] Fix double callback call when an error happens in the pipe.
Update transform is designed to fail when a specific string is passed.
It tests for that case and if the string matches, it calls the callback
with an appropriate error object but, instead of returning, it keeps
goin and calls the callback again, raising a 'write callback called
multiple times' error.

Add a return statement after calling the callback with error object
to avoid disruptive behavior.