Cleanup event handlers before return.
After the promisePipe returns, it leaves some event handlers installed
in the streams. If the user later wants to reuse the streams, the
handlers will be called out of proper contexts, which can lead to bad
code behavior for no clear sane reason.

We fix this by uninstalling all event handlers before the function
returns or when an error happens.

Notice that any prior event handler the user installed for the relevant
events will be lost after the function returns. Fixing this would
require a somewhat bigger code change, which seems not worth to cover
such corner case.